- tabs fixed - KDE stuff moved glib
