+++ src/glib/base/xmath.cpp
@@ -1350,24 +1350,24 @@ void TSvd::Wr() const {
 /////////////////////////////////////////////////
 // Histogram
 void THist::Add(const double& Val, const bool& OnlyInP) {
-	// get bucket number
+    // get bucket number
     const int BucketN = int(floor((Val - MnVal) / BucketSize));
-	if (OnlyInP) { 
-		// value should be inside
-		EAssert(MnVal <= Val && Val <= MxVal);
-		BucketV[BucketN]++;
-	} else {
-		// value does not need to be inside
-		if (BucketN < 0) {
-			BucketV[0]++;
-		} else if (BucketN < BucketV.Len()) {
-			BucketV[BucketN]++;
-		} else {
-			BucketV.Last()++;
-		}
-	}
-	// for computing percentage
-	Vals++;
+    if (OnlyInP) {
+        // value should be inside
+        EAssert(MnVal <= Val && Val <= MxVal);
+        BucketV[BucketN]++;
+    } else {
+        // value does not need to be inside
+        if (BucketN < 0) {
+            BucketV[0]++;
+        } else if (BucketN < BucketV.Len()) {
+            BucketV[BucketN]++;
+        } else {
+            BucketV.Last()++;
+        }
+    }
+    // for computing percentage
+    Vals++;
 }
 
 void THist::SaveStat(const TStr& ValNm, TSOut& FOut) const {
@@ -1386,26 +1386,26 @@ void THist::SaveStat(const TStr& ValNm, TSOut& FOut) const {
 /////////////////////////////////////////////////
 // Statistics
 void TStatFun::ChiSquare(const TFltV& OutValVX, const TFltV& OutValVY, const TInt& Df,
-		TFlt& Chi2, TFlt& P) {
-	Chi2 = 0.0;
-	P = 1.0;
-	EAssertR(OutValVX.Len() == OutValVY.Len(), "TChiSquare: histogram dimensions do not match!");
-	// http://www.itl.nist.gov/div898/software/dataplot/refman1/auxillar/chi2samp.htm
-	double SumR = TLinAlg::SumVec(OutValVX);
-	double SumS = TLinAlg::SumVec(OutValVY);
-	// Do nothing if zero histogram is detected
-	if (SumR <= 0.0 || SumS <= 0.0) { return; }
-	double K1 = TMath::Sqrt(SumS / SumR);
-	double K2 = 1.0 / K1;
-	for (int ValN = 0; ValN < OutValVX.Len(); ValN++) {
-		double Ri = OutValVX[ValN];
-		double Si = OutValVY[ValN];
-		double RpS = Ri + Si;
-		if (RpS > 0) {
-			Chi2 += TMath::Sqr(K1 * Ri - K2 * Si) / RpS;
-		}
-	}
-	if (Chi2 != 0.0) {
-		P = TSpecFunc::GammaQ(0.5*(Df), 0.5*(Chi2));
-	}
+    TFlt& Chi2, TFlt& P) {
+    Chi2 = 0.0;
+    P = 1.0;
+    EAssertR(OutValVX.Len() == OutValVY.Len(), "TChiSquare: histogram dimensions do not match!");
+    // http://www.itl.nist.gov/div898/software/dataplot/refman1/auxillar/chi2samp.htm
+    double SumR = TLinAlg::SumVec(OutValVX);
+    double SumS = TLinAlg::SumVec(OutValVY);
+    // Do nothing if zero histogram is detected
+    if (SumR <= 0.0 || SumS <= 0.0) { return; }
+    double K1 = TMath::Sqrt(SumS / SumR);
+    double K2 = 1.0 / K1;
+    for (int ValN = 0; ValN < OutValVX.Len(); ValN++) {
+        double Ri = OutValVX[ValN];
+        double Si = OutValVY[ValN];
+        double RpS = Ri + Si;
+        if (RpS > 0) {
+            Chi2 += TMath::Sqr(K1 * Ri - K2 * Si) / RpS;
+        }
+    }
+    if (Chi2 != 0.0) {
+        P = TSpecFunc::GammaQ(0.5*(Df), 0.5*(Chi2));
+    }
 }
