+++ src/glib/mine/mine.h
@@ -82,4 +82,7 @@
 // non-negative matrix factorization
 #include "nmf.h"
 
+// kernel-density-estimation
+#include "kerneldensity.h"
+
 #endif
